LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-strict_dynamic_worker-importScripts.https-expected.txt


PASS `importScripts(...)` is allowed by 'strict-dynamic'
